includes function for updating table rows

+++ functions.py
@@ -71,7 +71,6 @@ def write_table(u_service, u_login, c_password): #function that stores data in t
     db, mycursor = connect_database() #connects to database
     mycursor.execute("INSERT INTO PasswordsTable (service, login, password) VALUES (%s, %s, %s)",(u_service, u_login, c_password)) #SQL script to insert new row into table
     db.commit() #commmit changes in table
-    print("\n---------- Password successfully inserted into table! ----------")
 
 def check_if_registered(u_service): #function that checks if some string is already present in database
     db, mycursor = connect_database() #connects to database
@@ -81,22 +80,81 @@ def check_if_registered(u_service): #function that checks if some string is alre
     select_query = f"SELECT * FROM PasswordsTable WHERE service = '{u_service}'" #tries to get stored info for this service
     mycursor.execute(select_query)
     if len(mycursor.fetchall()) > 0:
-        print("\n'{}' already has a password stored in the database. \n".format(u_service))
-        print("\n---------- Please select a new option below ----------")
         service_registered = True
     
     return service_registered
 
 def delete_table_row(u_service): #function that deletes one row from table
+    service_found = check_if_registered(u_service) #check if this service is already registered in database
+    
+    if service_found == False:
+        print("\n'{}' does not have any register stored in the database. \n".format(u_service))
+        print("\n---------- Please select a new option below ----------")
+        return None
+
     db, mycursor = connect_database() #connects to database
     delete_query = f"DELETE FROM PasswordsTable WHERE service = '{u_service}'" #query to delete rows where the service is the one showed
     mycursor.execute(delete_query)
     db.commit()
-    #criar o código
 
-def change_table_row(): #function that updates one row from table
-    print("Hello")
-    #criar o código
+def change_table_row(u_service): #function that updates one row from table
+    service_found = check_if_registered(u_service) #check if this service is already registered in database
+    
+    if service_found == False:
+        print("\n'{}' does not have any register stored in the database. \n".format(u_service))
+        print("\n---------- Please select a new option below ----------")
+        return None
+
+    db, mycursor = connect_database() #connects to database
+    select_query = f"SELECT login, password FROM PasswordsTable WHERE service = '{u_service}'" #retrieves stored password and login for this service
+    mycursor.execute(select_query)
+    records = mycursor.fetchall()
+
+    for row in records:
+        u_login = row[0]
+        str_password = decrypt_string(row[1])
+
+    change_login = ""
+    change_password = ""
+
+    print("\n---------- Let's update {}'s records! ----------\n".format(u_service))
+    while change_login != "y" and change_login != "n":
+        try:
+            change_login = input("Would you like to change your login info? (y/n): ")
+        except ValueError:
+            continue
+    
+    if change_login == "y": #if a new login will be added
+            u_login = input("What is your login (or email)? ") #login info
+
+    while change_password != "y" and change_password != "n":
+        try:
+            change_password = input("Would you like to change your password info? (y/n): ")
+        except ValueError:
+            continue
+
+    if change_password == "y": #if a new password will be added
+        secure_password = ""
+        while secure_password != "y" and secure_password != "n":
+            try:
+                secure_password = input("Would you like to generate a new secure password? (y/n): ")
+            except ValueError:
+                continue
+        if secure_password == "y":
+            str_password = generate_password() #calls function to generate the password
+        elif secure_password == "n":
+            str_password = input("What is your new password? ")
+
+    if change_login == "n" and change_password == "n":
+        print("\n---------- Nothing to be updated! ----------")
+        return None        
+    
+    u_password = encrypt_string(str_password) #encrypts password
+    
+    delete_table_row(u_service) #deletes outdated row in table
+    write_table(u_service, u_login, u_password) #writes new row with updated password
+    db.commit() #commits changes into DB
+    print("\n---------- {} register successfully updated! ----------\n".format(u_service))
 
 def show_password(u_service): #function that returns one password from the table
     db, mycursor = connect_database() #connects to database
